RavenDB-23159 Rename bits to partials
