+++ src/Raven.Studio/typescript/components/common/virtualTable/VirtualTableWithLazyLoading.tsx
@@ -2,7 +2,7 @@ import "./VirtualTable.scss";
 import { flexRender } from "@tanstack/react-table";
 import { ClassNameProps } from "components/models/common";
 import { virtualTableConstants } from "components/common/virtualTable/utils/virtualTableConstants";
-import VirtualTableBodyWrapper, { VirtualTableBodyWrapperProps } from "./bits/VirtualTableBodyWrapper";
+import VirtualTableBodyWrapper, { VirtualTableBodyWrapperProps } from "./partials/VirtualTableBodyWrapper";
 import classNames from "classnames";
 
 // Use it along with useVirtualTableWithLazyLoading hook
