+++ src/Raven.Studio/typescript/components/common/virtualTable/partials/VirtualTableHead.tsx
