+++ src/Raven.Studio/typescript/components/common/virtualTable/partials/VirtualTableState.tsx
